working script copy-pasted

+++ script/codeship_build.sh
@@ -1,5 +1,4 @@
 # NOTE this script MIGHT have to be copy/pasted into the "setup commands" of the codeship project configuration.
-
 # We support all major Ruby versions. Please see our documentation for a full list.
 # https://documentation.codeship.com/basic/languages-frameworks/ruby/
 # If `.ruby-version` file is present, RVM will set Ruby to the declared version.
@@ -9,10 +8,7 @@ if [ -f .ruby-version ]; then rvm use $(cat .ruby-version) --install; fi
 nvm install 10.23.0
 npm install --global npm@"$(jq -r '.engines.npm' package.json)"
 mkdir -p node_modules/.bin
-if [ ! -e "node_modules/.bin/nodejs" ]; then
-  echo "creating link"
-  ln -s $(which node) node_modules/.bin/nodejs
-fi
+if [ ! -e "node_modules/.bin/nodejs" ]; then ln -s $(which node) node_modules/.bin/nodejs; fi
 echo "NODE: $(nodejs --version)"
 rvm use 2.6.3 --install
 gem install bundler
